comment resolution 2
